+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 For instructions on upgrading to newer versions, visit
 [mongoid.org](http://mongoid.org/en/mongoid/docs/upgrading.html).
 
-## 3.1.0 (branch: master)
+## 3.1.0
 
 ### New Features